chore: move form elements
