+++ src/components/Form/FileInput.field.tsx
@@ -1,5 +1,5 @@
 import type { FieldProps } from './types'
-import { FileInput } from '../FileInput/FileInput'
+import { FileInput } from './FileInput/FileInput'
 
 export const FileInputField = ({ input, ...rest }: FieldProps) => (
   <FileInput {...rest} onFilesChange={(files) => input.onChange(files)} />
