+++ src/components/Form/FileInput/UppyConfig.ts
